+++ ydb/core/formats/arrow/accessor/abstract/accessor.cpp
@@ -12,7 +12,8 @@
 #include <ydb/library/formats/arrow/switch/switch_type.h>
 #include <ydb/core/scheme_types/scheme_type_info.h>
 #include <contrib/libs/apache/arrow/cpp/src/arrow/array/array_primitive.h>
-#include <ydb/library/formats/arrow/scalar/serialization.h>
+
+#include <valarray>
 namespace NKikimr::NArrow::NAccessor {
 
 void IChunkedArray::TReader::AppendPositionTo(arrow::ArrayBuilder& builder, const ui64 position, ui64* recordSize) const {
@@ -213,65 +214,6 @@ TString IChunkedArray::TFullDataAddress::DebugString(const ui64 position, const 
 
 void IChunkedArray::TLocalDataAddress::Reallocate() {
     Array = NArrow::ReallocateArray(Array);
-}
-using TSizeType = ui32;
-constexpr char NullTMinMaxFlag = '0'; 
-
-TMinMax TMinMax::FromBinaryString(std::string_view data, const std::shared_ptr<arrow::DataType> &fieldType) {
-    if (data[0] == NullTMinMaxFlag) {
-        return MakeNull(fieldType);
-    }
-    std::vector<std::shared_ptr<arrow::Scalar>> fields;
-    ui64 offset = 1;
-    auto readNext = [&] {
-        TSizeType size = 0;
-        AFL_VERIFY(offset <= data.size() && sizeof(TSizeType) <= data.size() - offset )("details",
-                                                               Sprintf("out of bounds read, data.size(): %i, current_offset: %i, read size: %i",
-                                                                   data.size(), offset, sizeof(TSizeType)));
-        memcpy(&size, data.data() + offset, sizeof(size));
-        offset += sizeof(TSizeType);
-        AFL_VERIFY(offset <= data.size() && size <= data.size() - offset)(
-                                                  "details", Sprintf("out of bounds read, data.size(): %i, current_offset: %i, read size: %i",
-                                                                 data.size(), offset, size));
-        auto res =
-            NArrow::NScalar::TSerializer::DeserializeFromStringWithPayload(TStringBuf{ data.data() + offset, data.data() + offset + size }, fieldType)
-                .DetachResult();
-        offset += size;
-        return res;
-    };
-    fields.push_back(readNext());
-    fields.push_back(readNext());
-        return 
-            *arrow::StructScalar::Make(fields, 
-            {"min", "max"}).ValueOrDie();
-}
-
-std::string TMinMax::ToBinaryString() const {
-    if (IsNull()) {
-        return std::string{NullTMinMaxFlag};
-    }
-    std::string res;
-    res.push_back(NullTMinMaxFlag + 1);
-    TString minSerialized = NArrow::NScalar::TSerializer::SerializePayloadToString(Min()).DetachResult();
-    TString maxSerialized = NArrow::NScalar::TSerializer::SerializePayloadToString(Max()).DetachResult();
-    auto writeNext = [&](TStringBuf data) {
-        TSizeType dataSize = data.size();
-        ui64 resSize = res.size();
-        res.resize(resSize + sizeof(TSizeType));
-        memcpy(res.data() + resSize, &dataSize, sizeof(TSizeType));
-        res.append(data);
-    };
-    writeNext(minSerialized);
-    writeNext(maxSerialized);
-    return res;
-
-}
-
-NJson::TJsonValue TMinMax::Json() const {
-    NJson::TJsonValue json;
-    json.InsertValue("min", Min()->ToString());
-    json.InsertValue("max", Max()->ToString());
-    return json;
 }
 
 }   // namespace NKikimr::NArrow::NAccessor
